nits(nvidia/query): move xid device handlers
