+++ components/accelerator/nvidia/query/nvml/gpm.go
@@ -1,0 +1,1 @@
+package nvml
