+++ components/accelerator/nvidia/query/nvml/nvml.go
@@ -141,7 +141,7 @@ func NewInstance(ctx context.Context) (Instance, error) {
 		nvmlExists:    nvmlExists,
 		nvmlExistsMsg: nvmlExistsMsg,
 		eventSet:      eventSet,
-		eventMask:     defaultEventMask,
+		eventMask:     defaultXidEventMask,
 		eventCh:       make(chan *XidEvent, 100),
 	}, nil
 }
@@ -149,8 +149,6 @@ func NewInstance(ctx context.Context) (Instance, error) {
 func (inst *instance) NVMLExists() bool {
 	return inst.nvmlExists
 }
-
-const defaultEventMask = uint64(nvml.EventTypeXidCriticalError | nvml.EventTypeDoubleBitEccError | nvml.EventTypeSingleBitEccError)
 
 // Starts an NVML instance and starts polling for XID events.
 // ref. https://github.com/NVIDIA/k8s-device-plugin/blob/main/internal/rm/health.go
@@ -231,70 +229,6 @@ func (inst *instance) Start() error {
 	go inst.pollXidEvents()
 
 	return nil
-}
-
-// ref. https://docs.nvidia.com/deploy/nvml-api/group__nvmlEvents.html#group__nvmlEvents
-func (inst *instance) pollXidEvents() {
-	log.Logger.Debugw("polling xid events")
-	for {
-		select {
-		case <-inst.rootCtx.Done():
-			return
-		default:
-		}
-
-		// waits 5 seconds
-		// ref. https://docs.nvidia.com/deploy/nvml-api/group__nvmlEvents.html#group__nvmlEvents
-		e, ret := inst.eventSet.Wait(5000)
-
-		if ret == nvml.ERROR_TIMEOUT {
-			log.Logger.Debugw("no event found in wait (timeout) -- retrying...", "error", nvml.ErrorString(ret))
-			continue
-		}
-
-		if ret != nvml.SUCCESS {
-			select {
-			case <-inst.rootCtx.Done():
-				return
-
-			case inst.eventCh <- &XidEvent{
-				Message: "event set wait returned non-success",
-				Error:   fmt.Errorf("event set wait failed: %v", nvml.ErrorString(ret)),
-			}:
-				log.Logger.Debugw("event set wait failure notified", "error", nvml.ErrorString(ret))
-			}
-
-			continue
-		}
-
-		xid := e.EventData
-
-		var xidDetail *nvidia_query_xid.Detail
-		msg := "received event but xid unknown"
-		if xid > 0 {
-			var ok bool
-			xidDetail, ok = nvidia_query_xid.GetDetail(int(xid))
-			if ok {
-				msg = "received event with a known xid"
-			}
-		}
-
-		event := &XidEvent{
-			EventType: e.EventType,
-
-			Xid:              xid,
-			XidCriticalError: e.EventType == nvml.EventTypeXidCriticalError,
-
-			Detail: xidDetail,
-
-			Message: msg,
-		}
-		select {
-		case <-inst.rootCtx.Done():
-			return
-		case inst.eventCh <- event:
-		}
-	}
 }
 
 func (inst *instance) RecvXidEvents() <-chan *XidEvent {
